+++ internal/ast/print.go
@@ -14,7 +14,7 @@ func Fprint(w io.Writer, prog *Program) error {
 	}
 	p := &printer{w: w}
 	p.line(`algoritmo "%s"`, prog.Name)
-	p.printDecls(prog.Consts, prog.Globals)
+	p.printDecls(prog.Consts, prog.Types, prog.Globals)
 	for _, sub := range prog.Subs {
 		p.line("")
 		p.printSub(sub)
@@ -50,8 +50,8 @@ func (p *printer) line(format string, args ...any) {
 	_, p.err = fmt.Fprintln(p.w)
 }
 
-func (p *printer) printDecls(consts []ConstDecl, decls []VarDecl) {
-	if len(consts) == 0 && len(decls) == 0 {
+func (p *printer) printDecls(consts []ConstDecl, types []TypeDecl, decls []VarDecl) {
+	if len(consts) == 0 && len(types) == 0 && len(decls) == 0 {
 		return
 	}
 	if len(consts) != 0 {
@@ -59,6 +59,14 @@ func (p *printer) printDecls(consts []ConstDecl, decls []VarDecl) {
 		p.indent++
 		for _, d := range consts {
 			p.line("%s = %s", d.Name.Text, exprString(d.Value))
+		}
+		p.indent--
+	}
+	if len(types) != 0 {
+		p.line("tipo")
+		p.indent++
+		for _, d := range types {
+			p.line("%s = %s", d.Name.Text, typeString(d.Type))
 		}
 		p.indent--
 	}
@@ -78,7 +86,7 @@ func (p *printer) printSub(sub Subprogram) {
 	switch s := sub.(type) {
 	case *ProcedureDecl:
 		p.line("procedimento %s(%s)", s.Name.Text, paramsString(s.Params))
-		p.printDecls(s.Consts, s.Locals)
+		p.printDecls(s.Consts, s.Types, s.Locals)
 		p.line("inicio")
 		p.indent++
 		p.printStmts(s.Body)
@@ -86,7 +94,7 @@ func (p *printer) printSub(sub Subprogram) {
 		p.line("fimprocedimento")
 	case *FunctionDecl:
 		p.line("funcao %s(%s): %s", s.Name.Text, paramsString(s.Params), typeString(s.Return))
-		p.printDecls(s.Consts, s.Locals)
+		p.printDecls(s.Consts, s.Types, s.Locals)
 		p.line("inicio")
 		p.indent++
 		p.printStmts(s.Body)